[Intl] Renamed StructuredBundleReader[Interface] to BundleEntryReader[Interface]
