+++ src/Symfony/Component/Intl/CHANGELOG.md
@@ -12,9 +12,11 @@ CHANGELOG
    names when fallback (e.g. "en_GB"->"en") is possible anyway. This reduced
    the Resources/ directory file size of the Icu 1.2.x branch from 14M to 12M at
    the time of this writing
- * `StructuredBundleReader` now follows aliases when looking for fallback locales
  * [BC BREAK] a new method `getLocaleAliases()` was added to `LocaleBundleInterface`
+ * deprecated `StructuredBundleReader` and `StructuredBundleReaderInterface` in
+   favor of `BundleEntryReader` and `BundleEntryReaderInterface`
+ * `BundleEntryReader` now follows aliases when looking for fallback locales
  * [BC BREAK] changed default value of the argument `$fallback` in the protected
    method `AbstractBundle::readEntry()` to `true` in order to be consistent with
-   the proxied `StructuredBundleReaderInterface::readEntry()` method
+   the proxied `BundleEntryReaderInterface::readEntry()` method
  * deprecated `BundleCompiler` in favor of `GenrbBundleCompiler`
