+++ src/Symfony/Component/Intl/ResourceBundle/AbstractBundle.php
@@ -46,7 +46,7 @@ abstract class AbstractBundle implements ResourceBundleInterface
     }
 
     /**
-     * Proxy method for {@link StructuredBundleReaderInterface#read}.
+     * Proxy method for {@link BundleEntryReaderInterface#read}.
      */
     protected function read($locale)
     {
@@ -54,7 +54,7 @@ abstract class AbstractBundle implements ResourceBundleInterface
     }
 
     /**
-     * Proxy method for {@link StructuredBundleReaderInterface#readEntry}.
+     * Proxy method for {@link BundleEntryReaderInterface#readEntry}.
      */
     protected function readEntry($locale, array $indices, $fallback = true)
     {
